+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the TailSocks project will be documented in this file. This project follows the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) standard.
 
-## [4.2.0] - Unreleased
+## [4.1.1] - 2026-09-07
 
 ### Added